Attempt to fix Windows bug #2

+++ include/dynd/kernels/assignment_kernels.hpp
@@ -165,7 +165,7 @@ namespace nd {
 
       DYND_TRACE_ASSIGNMENT(d, dst_type, s, src0_type);
 
-      *reinterpret_cast<dst_type *>(dst) = static_cast<float>(s);
+      *reinterpret_cast<dst_type *>(dst) = static_cast<double>(s);
     }
   };
